+++ src/test/setup.ts
@@ -5,6 +5,7 @@ import { afterEach } from 'vitest'
 afterEach(() => {
   cleanup()
   window.localStorage.clear()
+  window.history.replaceState({}, '', '/')
 })
 
 Object.defineProperty(window, 'matchMedia', {